Refactor: 공통 타입을 따로 빼주는것이 아닌 CurrentWeather, FiveDayWeather 타입에 각각 구현해주는 방향으로 수정
