--- WeatherForecast/WeatherForecast/Model/Clouds.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-struct Clouds: Decodable {
-    let allOver: Int
-    
-    enum CodingKeys: String, CodingKey {
-        case allOver = "all"
-    }
-}
