+++ WeatherForecast/WeatherForecast/Model/CurrentWeather.swift
@@ -4,11 +4,13 @@ struct CurrentWeather: Decodable {
     let locationCoordinate: LocationCoordinate
     let weather: [Weather]
     let base: String
-    let main: CurrentWeatherMain
+    let main: Main
     let wind: Wind
     let clouds: Clouds
+    let rain: Rain
+    let snow: Snow
     let coordinatedUniversalTime: Date
-    let sys: CurrentWeatherSys
+    let sys: Sys
     let timezone: Int
     let cityID: Int
     let cityName: String
@@ -16,13 +18,36 @@ struct CurrentWeather: Decodable {
     
     enum CodingKeys: String, CodingKey {
         case locationCoordinate = "coord"
-        case weather, base, main, wind, clouds, sys, timezone, cod
+        case weather, base, main, wind, clouds, rain, snow, sys, timezone, cod
         case coordinatedUniversalTime = "dt"
         case cityID = "id"
         case cityName = "name"
     }
     
-    struct CurrentWeatherMain: Decodable {
+    struct LocationCoordinate: Decodable {
+        let longitude: Double
+        let latitude: Double
+        
+        enum CodingKeys: String, CodingKey {
+            case longitude = "lon"
+            case latitude = "lat"
+        }
+    }
+    
+    struct Weather: Decodable {
+        let stateID: Int
+        let main: String
+        let description: String
+        let iconID: String
+        
+        enum CodingKeys: String, CodingKey {
+            case stateID = "id"
+            case main, description
+            case iconID = "icon"
+        }
+    }
+    
+    struct Main: Decodable {
         let temperature: Double
         let feelsLike: Double
         let minTemperature: Double
@@ -59,7 +84,45 @@ struct CurrentWeather: Decodable {
         }
     }
     
-    struct CurrentWeatherSys: Decodable {
+    struct Wind: Decodable {
+        let speed: Double
+        let degree: Int
+        
+        enum CodingKeys: String, CodingKey {
+            case speed
+            case degree = "deg"
+        }
+    }
+    
+    struct Clouds: Decodable {
+        let allOver: Int
+        
+        enum CodingKeys: String, CodingKey {
+            case allOver = "all"
+        }
+    }
+    
+    struct Rain: Decodable {
+        let oneHour: Double
+        let threeHour: Double
+        
+        enum CodingKeys: String, CodingKey {
+            case oneHour = "1h"
+            case threeHour = "3h"
+        }
+    }
+    
+    struct Snow: Decodable {
+        let oneHour: Double
+        let threeHour: Double
+        
+        enum CodingKeys: String, CodingKey {
+            case oneHour = "1h"
+            case threeHour = "3h"
+        }
+    }
+    
+    struct Sys: Decodable {
         let type: Int
         let id: Int
         let message: Double
