--- WeatherForecast/WeatherForecast/Model/LocationCoordinate.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct LocationCoordinate: Decodable {
-    let longitude: Double
-    let latitude: Double
-    
-    enum CodingKeys: String, CodingKey {
-        case longitude = "lon"
-        case latitude = "lat"
-    }
-}
