--- WeatherForecast/WeatherForecast/Model/Wind.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct Wind: Decodable {
-    let speed: Double
-    let degree: Int
-    
-    enum CodingKeys: String, CodingKey {
-        case speed
-        case degree = "deg"
-    }
-}
